// FoodDiary.Web.Client/src/app/shared/api/sdk/generated/api/recipes.service.ts
export interface GetRecipesRequestParams {
    version: string;
    page?: number;
    limit?: number;
    search?: string;
    includePublic?: boolean;
    category?: string;
    maxTotalTime?: number;
    caloriesFrom?: number;
    caloriesTo?: number;
    hasImage?: boolean;
}